// examples/showcase/project-tracker/src/types/projects.ts
/**
 * ObjectQL
 * Copyright (c) 2026-present ObjectStack Inc.
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */

// Auto-generated by ObjectQL. DO NOT EDIT.
import { ObjectDoc } from '@objectql/types';

export interface Projects extends ObjectDoc {
    /**
     * Name
     */
    name: string;
    /**
     * Status
     */
    status?: string;
    /**
     * Priority
     */
    priority?: string;
    /**
     * Description
     */
    description?: string;
    /**
     * Owner
     */
    owner?: string;
    /**
     * Budget
     */
    budget?: number;
    /**
     * Start Date
     */
    start_date?: Date | string;
    /**
     * End Date
     */
    end_date?: Date | string;
    /**
     * Approved By
     */
    approved_by?: string;
    /**
     * Approved At
     */
    approved_at?: Date | string;
    /**
     * Approval Comment
     */
    approval_comment?: string;
}